ci: Jenkins Playwright Tests
This commit removes the build via test method that was implemented to reduce execution time. The idea being go test also performs a build then we should not be calling build before test, or at all for that matter.

We are not executing unit tests in this pipeline but will need to build and install all necessary storj packages.